+++ CHANGELOG.md
@@ -38,10 +38,12 @@
 - PR #414 Add weights to wjaccrd.py
 - PR #436 Fix Skip Test Functionality
 - PR #438 Fix versions of packages in build script and conda yml
+- PR #441 Import cudf_cpp.pxd instead of duplicating cudf definitions.
+- PR #441 Removed redundant definitions of python dictionaries and functions.
 - PR #442 Updated versions in conda environments.
 - PR #442 Added except + to cython bindings to C(++) functions.
 - PR #443 Fix accuracy loss issue for snmg pagerank
-
+- PR #444 Fix warnings in strongly connected components
 
 # cuGraph 0.8.0 (27 June 2019)